+++ SERVE.sh
@@ -1,1 +1,1 @@
-python -m SimpleHTTPServer 8000
+python3 -m http.server 8000