+++ docs/CHANGE_LOG.md
@@ -3,6 +3,7 @@
 ## Version 2.9.3:
 - Added new setting "minutesBetweenSectionsInViews" (defaults to 30), which states the minutes that should be used between headers/rows in all views for time displays.
 - Added new setting "timelineViewDefaultAxis" (defaults to "group"), which states the default axis the Timeline view should use.
+- Updated SHARE.sh to use Python 3 for running a simple HTTP Server.
 - 
 
 <br>